+++ src/index.js
@@ -1,9 +1,18 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import Head from './components/Header/header';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
+import {
+  BrowserRouter,
+  Routes,
+  Route,
+} from "react-router-dom";
 
-
-ReactDOM.render(
-    <Head/>,
-    document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </React.StrictMode>
 );
